change radius from int to float

+++ Shared/Shared/Ball.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <glm/gtx/string_cast.hpp>
 
-Ball::Ball(vec3 position, vec3 velocity, int id, int radius) : GameObject(position, velocity, id, radius) {
+Ball::Ball(vec3 position, vec3 velocity, int id, float radius) : GameObject(position, velocity, id, radius) {
 }
 
 int Ball::getGameObjectType() const {
