+++ Shared/Shared/Paddle.cpp
@@ -1,6 +1,6 @@
 #include "Paddle.h"
 
-Paddle::Paddle (vec3 position, vec3 velocity, int id, int radius, int lifespan) : GameObject(position, velocity, id, radius) {
+Paddle::Paddle (vec3 position, vec3 velocity, int id, float radius, int lifespan) : GameObject(position, velocity, id, radius) {
 	this->lifespan = lifespan;
 }
